fixed some modules, added testbench module
INC and DEC had some errors. INC checked if the input equals 1, not if all bits equal 1. DEC would set output to 1 if input was 0, not set all inputs to 1. I have replaced the logic with the built in + and - operators which check for overflow (I checked with the testbench). Also I have commented lines with what I believe would be correct code for explicit over/underflow checking in the INC/DEC modules.

Test bench might be a bit overkill..